Moves SDK-related sweeper functions to separate package
